+++ packages/components/src/components/navigation-item/navigation-item.stories.ts
@@ -9,8 +9,12 @@ const { generateTemplate } = storybookTemplate('sd-navigation-item');
 /**
  * Used to facilitate seamless page transitions and helps users orient themselves within the application.
  *
+ * **Related components**:
+ * - [sd-header](?path=/docs/components-sd-header--docs)
+ *
  * **Related templates**:
- * - [Dropdown with Navigation Items](?path=/docs/templates-dropdown-with-navigation-items--docs)
+ * - [Dropdown with Navigation Items](?path=/docs/templates-dropdown--docs#dropdown-with-navigation-items)
+ * - [Navigation Item with Badge](?path=/docs/templates-badge--docs#navigation%20item%20with%20badge)
  */
 
 export default {
@@ -35,24 +39,11 @@ export const Default = {
 };
 
 /**
- * Use the `vertical` attribute to set the axis of the navigation-item.
- * - default: horizontal navigation for headers
- * - vertical: vertical navigation for e.g. drawers
- */
-
-export const Orientation = {
-  render: () =>
-    html`<sd-navigation-item>Horizontal Navigation</sd-navigation-item>
-      <sd-navigation-item vertical>Vertical Navigation</sd-navigation-item>`
-};
-
-/**
- * Use the `size` attribute to change the font size of the navigation item.
- * - `sm`: used for 3rd level navigation
+ * Use the `size` attribute to change the font size of the navigation item:
+ * - `sm`: Used for 3rd level navigation
  * - `base` (default)
- * - `lg`: used for 2nd level navigation
+ * - `lg`: Used for 2nd level navigation
  */
-
 export const Size = {
   render: () => html`
     <div class="flex flex-col gap-6">
@@ -64,22 +55,30 @@ export const Size = {
 };
 
 /**
- * Use the attribute `href` to change the navigation item to a link.
- * - `href` attribute - The URL to navigate to.
- * - `target` attribute - The target of the link. Can assume the values `_blank`, `_parent`, `_self`, or `_top`.
- * - `download` attribute - The filename to download the link as.
+ * Use the `vertical` attribute to set the axis of the navigation-item.
  */
 
-export const Link = {
+export const Orientation = {
+  render: () =>
+    html`<sd-navigation-item>Horizontal Navigation</sd-navigation-item>
+      <sd-navigation-item vertical>Vertical Navigation</sd-navigation-item>`
+};
+
+/**
+ * - Use the `href`attribute to change the navigation item to a link. Allows to set the URL to navigate to.
+ * - Use the `target` attribute to specify where to open the link. Can assume the values `_blank`, `_parent`, `_self`, or `_top`.
+ * - Use the `download` attribute to tell the browser to download the linked file as this filename.
+ */
+export const AsLink = {
+  name: 'As Link',
   render: () => html`
     <sd-navigation-item href="https://www.union-investment.de/" target="_blank">Link</sd-navigation-item>
   `
 };
 
 /**
- * Use the `current` attribute to change the navigation item to a current state and and make it bold.
+ * Use the `current` attribute to change the navigation item to a current state and make it bold.
  */
-
 export const Current = {
   render: () =>
     html`<div class="flex flex-col gap-6">
@@ -98,7 +97,8 @@ export const Disabled = {
 
 /**
  * Use the `divider` attribute to add a divider above the navigation item.
- * - Only works with `vertical` attribute.
+ *
+ * __Hint:__ Only works with `vertical` attribute.
  */
 
 export const Divider = {
@@ -107,6 +107,8 @@ export const Divider = {
 
 /**
  * Use the `chevron` attribute to add a chevron to the navigation item.
+ *
+ * __Hints:__
  * - Only works with `vertical` attribute.
  * - `Button` and `Link` variants show right facing chevron dependent on property.
  * - `Accordion` variant always show up / down chevron to reflect open state.
@@ -118,6 +120,8 @@ export const Chevron = {
 
 /**
  * Add `children` slot to the navigation item to create an accordion.
+ *
+ * __Hints:__
  * - Only works with `vertical` attribute.
  * - A `chevron` will be added regardless of the `chevron` attribute.
  * - The `open` attribute can be used to control the open state of the accordion.
@@ -133,10 +137,11 @@ export const Accordion = {
 };
 
 /**
- *
  * Use the `separated` attribute, to have more that only one action. It is possible to use it as a link and an accordion simultaneously.
  *
- * - It needs a `children slot` and an `href` attribute. `target` and `download` attributes are optional.
+ * __Hints:__
+ * - Only works with a `children slot` and an `href` attribute.
+ * - `target` and `download` attributes are optional.
  */
 export const Separated = {
   render: () =>
@@ -149,8 +154,9 @@ export const Separated = {
 };
 
 /**
- * Use the `description` to provide a description for the navigation item.
- * - Only works with `vertical` attribute.
+ * Use the `description` slot to provide a description for the navigation item.
+ *
+ * __Hint:__ Only works with `vertical` attribute.
  */
 export const Description = {
   render: () =>
@@ -164,7 +170,8 @@ export const Description = {
 
 /**
  * Use `indented` attribute to add padding to the left side.
- * - Only works with `vertical` attribute.
+ *
+ * __Hint:__ Only works with `vertical` attribute.
  */
 
 export const Indented = {
@@ -176,7 +183,8 @@ export const Indented = {
 
 /**
  * Use `relaxed` attribute to add padding to both sides.
- * - Only works with `vertical` attribute.
+ *
+ * __Hint:__ Only works with `vertical` attribute.
  */
 
 export const Relaxed = {