+++ packages/components/src/templates/drawer.stories.ts
@@ -4,10 +4,9 @@ import { html } from 'lit-html';
 /**
  * This example shows a drawer containing a navigation list.
  */
-
 export default {
   tags: ['!dev'],
-  title: 'Templates/Drawer with navigation',
+  title: 'Templates/Drawer',
   parameters: {
     design: {
       type: 'figma',
@@ -18,7 +17,7 @@ export default {
   decorators: [
     (story: any) =>
       html` <style>
-          #anchor--templates-drawer-with-navigation--default .innerZoomElementWrapper {
+          #anchor--templates-drawer--default .innerZoomElementWrapper {
             background-color: #ecf0f9;
             min-height: 750px;
           }
@@ -27,6 +26,9 @@ export default {
   ] as unknown
 };
 
+/**
+ * ### Tablet Navigation - Level 1
+ */
 export const Default = {
   name: 'Tablet Navigation - Level 1',
   render: () => html`