+++ packages/components/src/templates/map-marker.stories.ts
@@ -1,9 +1,13 @@
 import '../solid-components';
 import { html } from 'lit-html';
 
+/**
+ * ```
+ * ```
+ */
 export default {
   tags: ['!dev'],
-  title: 'Templates/Map Marker with Image',
+  title: 'Templates/Map Marker',
   parameters: {
     chromatic: { disableSnapshot: true },
     design: {
@@ -14,14 +18,11 @@ export default {
 };
 
 /**
- *
- * ```
- * ```
+ * ### Map Maker with Images
  */
-
-export const Default = {
+export const MapMakerWithImages = {
   render: () =>
     html`<sd-map-marker state="default" variant="place">
-      <img src="./placeholders/images/bild.png" alt="The Bild logo" class="h-8 w-8 -mt-4" />
+      <img src="images/ui-brand-mark.png" alt="Brand mark of Union Investment" class="h-8 w-8 -mt-4" />
     </sd-map-marker>`
 };