+++ packages/components/src/templates/mark.stories.ts
@@ -2,10 +2,9 @@ import '../solid-components';
 import { html } from 'lit-html';
 
 /**
- * **Accessibility:**
+ * __Accessibility Hint:__
  * For the green accent color in the mark, ensure a minimum font size of 18.67 pixels when bolded and 24 pixels when in regular weight.
  */
-
 export default {
   tags: ['!dev'],
   title: 'Templates/Mark',