+++ packages/components/src/templates/teaser-media.stories.ts
@@ -1,6 +1,10 @@
 import '../solid-components';
 import { html } from 'lit-html';
 
+/**
+ * ```
+ * ```
+ */
 export default {
   tags: ['!dev'],
   title: 'Templates/Teaser Media',
@@ -19,9 +23,9 @@ export const TeaserMediaWithLink = {
       <sd-teaser-media variant="gradient-dark">
         <div slot="media" class="relative">
           <img
-            class="w-full aspect-ratio "
-            src="./placeholders/images/skyscraper.jpg"
-            alt="A tall glass skyscraper viewed from below"
+            class="w-full aspect-ratio"
+            src="./placeholders/images/coins.jpg"
+            alt="Close-up of hands stacking coins into small piles on a table, suggesting financial planning or saving."
           />
         </div>
         <div slot="meta" class="meta-info">
@@ -42,8 +46,8 @@ export const TeaserMediaWithLink = {
         <img
           slot="media"
           class="w-full"
-          src="./placeholders/images/meeting.jpg"
-          alt="Three people in a business meeting talking to each other"
+          src="./placeholders/images/collaboration.jpg"
+          alt="Two men in business attire sitting on a sofa in a modern office, smiling and looking at a laptop screen."
         />
         <h3 slot="headline" class="sd-headline sd-headline--inverted">Your contact person</h3>
         <div class="flex flex-col gap-4">
@@ -60,8 +64,8 @@ export const TeaserMediaWithLink = {
         <img
           slot="media"
           class="w-full aspect-ratio"
-          src="./placeholders/images/flags.jpg"
-          alt="The U.S. and EU flags placed together"
+          src="./placeholders/images/skyline.jpg"
+          alt="City skyline of Frankfurt by night in christmas time."
         />
         <h3 slot="headline" class="sd-headline sd-headline--size-base">USA or Europe? It depends on the mix</h3>
         <div class="flex flex-col gap-4">
@@ -97,8 +101,8 @@ export const TeaserMediaOverrides = {
         <div slot="media" class="relative">
           <img
             class="w-full aspect-ratio"
-            src="./placeholders/images/coffeebreak.jpg"
-            alt="Two people sitting at a table, having a coffee in an office"
+            src="./placeholders/images/coffeeshop.jpg"
+            alt="People sitting at a table, having a coffee in a coffeeshop."
           />
         </div>
         <h3 slot="headline" class="sd-headline sd-headline--inverted">Gender</h3>
@@ -106,4 +110,45 @@ export const TeaserMediaOverrides = {
           <p class="sd-paragraph sd-paragraph--inverted text-xl pb-24">We are actively promoting gender equality.</p>
         </div>
       </sd-teaser-media> `
+};
+
+export const TeaserMediaWithCopyright = {
+  name: 'Teaser Media with Copyright',
+  render: () => html`
+    <style>
+      .sd-copyright::after {
+        padding: 0;
+      }
+      .gradient-white.sd-copyright::after {
+        color: #000000;
+        text-shadow: none;
+      }
+    </style>
+    <div class="flex flex-col gap-12">
+      <sd-teaser-media variant="gradient-dark" class="max-w-[600px]">
+        <img
+          slot="media"
+          src="./placeholders/images/architecture.jpg"
+          class="aspect-video object-cover"
+          alt="Generic alt"
+        />
+        <h3 slot="headline">Headline Media Teaser</h3>
+        <div class="flex flex-col sd-copyright" style="--copyright: '© Union Investment 2024'">
+          <div slot="default" class="h-16">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>
+        </div>
+      </sd-teaser-media>
+      <sd-teaser-media variant="gradient-white" class="max-w-[600px]">
+        <img
+          slot="media"
+          src="./placeholders/images/architecture.jpg"
+          class="aspect-video object-cover"
+          alt="Generic alt"
+        />
+        <h3 slot="headline">Headline Media Teaser</h3>
+        <div class="flex flex-col sd-copyright gradient-white" style="--copyright: '© Union Investment 2024'">
+          <div slot="default" class="h-16">Lorem ipsum dolor sit amet, consectetur adipiscing elit.</div>
+        </div>
+      </sd-teaser-media>
+    </div>
+  `
 };